Implemented suggested changes.
Fixes #94
